requirements.txt
Flask==3.1.2
adafruit-circuitpython-dht==4.0.9
Adafruit-Blinka==8.66.0
RPi.GPIO==0.7.1
